+++ speech-to-text-disabled.js
@@ -1,0 +1,9 @@
+const EventEmitter = require("events");
+
+function Speech() {
+  var t = new EventEmitter();
+  return t
+}
+
+const sp = Speech();
+module.exports = sp;
